Modified hasMatch to use long document id

+++ contrib/src/main/java/org/lemurproject/galago/contrib/index/disk/BackgroundStatsReader.java
@@ -158,7 +158,7 @@ public class BackgroundStatsReader extends KeyValueReader implements AggregateIn
     }
 
     @Override
-    public boolean hasMatch(int identifier) {
+    public boolean hasMatch(long identifier) {
       throw new UnsupportedOperationException("Not supported yet.");
     }
